backend/scripts_automation/app/requirements.txt
# FastAPI and web server
fastapi>=0.95.0
pydantic>=1.10.7*
uvicorn>=0.21.1
# Database
sqlmodel>=0.0.8
sqlalchemy>=2.0.9
psycopg2-binary>=2.9.6
pymysql>=1.0.3
pymongo>=4.3.3
# Utilities
python-dotenv>=1.0.0
requests>=2.28.2
croniter>=1.4.1
asyncio>=3.4.3

uvicorn[standard]
pandas

apscheduler


python-multipart
scikit-learn
joblib
matplotlib
xgboost
Authlib
pyotp
alembic
pyotp 
aiosmtplib
msal
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
sentence-transformers
firebase-admin

# Database connectors
mysql-connector-python>=8.0.0
snowflake-connector-python>=3.0.0
boto3>=1.26.0
redis>=4.5.0
pymongo>=4.3.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
pymysql>=1.0.0

# Core dependencies

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
aiofiles>=0.8.0
cryptography>=37.0.0
gunicorn>=20.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
httpx>=0.24.0
pytest-cov>=3.0.0

# Utilities
python-dateutil>=2.8.2
email-validator>=1.1.3
Jinja2>=3.0.0
aiohttp>=3.8.0
requests>=2.28.0
tenacity>=8.0.0
structlog>=22.1.0


# Cloud Services
boto3>=1.34.0  # AWS services
azure-identity>=1.15.0  # Azure managed identity
azure-storage-blob>=12.19.0  # Azure blob storage
azure-keyvault-secrets>=4.7.0  # Azure key vault
google-cloud-storage>=2.14.0  # GCP storage




# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.1
requests>=2.31.0
urllib3>=2.1.0